Change Download return signature to be (string, error)
One usecase I would like to support is the ability to uncompress the
downloaded file and delete it. I don't believe this is the
responsibility of the client, so to support this I am returning the
filename of the downloaded file.

This will enable the following:

```
filename, err := mediafire.Download(url, root)
filePath := filepath.Join(root, filename)
err := archiver.Unarchive(filepath, filename)
...
```