refactor: Download now accepts an io.Writer instead of filename
Change Download method signature to accept a url and an io.Writer to
stream the content of the file to. This opens up the possibility to
handling _many_ different cases instead of being limited to only the
local filesystem. No example exists yet, but you should be able to now
write out the results to a bucket.